DEV: refactor PyFibreCLI class to envisage Application
