+++ pyfibre/cli/tests/test_click_app.py
@@ -4,14 +4,16 @@ from unittest import mock, TestCase
 
 from click.testing import CliRunner
 
-import pyfibre.cli.app
+import pyfibre.cli.pyfibre_cli
 from pyfibre.version import __version__
 
-from pyfibre.tests.dummy_classes import DummyPyFibreCLI
+
+def mock_run_pyfibre():
+    pass
 
 
 def mock_run_constructor(*args, **kwargs):
-    mock_pyfibre_app = mock.Mock(spec=pyfibre.cli.app)
+    mock_pyfibre_app = mock.Mock(spec=pyfibre.cli.pyfibre_cli)
     mock_pyfibre_app.run = lambda: None
 
 
@@ -19,39 +21,43 @@ class TestClickRun(TestCase):
 
     def test_click_cli_version(self):
         clirunner = CliRunner()
-        clirunner.invoke(pyfibre.cli.app.pyfibre,
+        clirunner.invoke(pyfibre.cli.pyfibre_cli.pyfibre,
                          args="--version")
 
     def test_click_gui_main(self):
 
-        with mock.patch('pyfibre.cli.app') as mock_run:
+        with mock.patch('pyfibre.cli.pyfibre_cli') as mock_run:
             mock_run.side_effect = mock_run_constructor
 
-            pyfibre.cli.app.pyfibre()
+            pyfibre.cli.pyfibre_cli.pyfibre()
 
             self.assertTrue(mock_run.pyfibre.called)
 
     def test_run(self):
-        with mock.patch('pyfibre.cli.app.PyFibreCLI') as mock_pyfibre:
-            mock_pyfibre.return_value = DummyPyFibreCLI()
-            pyfibre.cli.app.run(
-                file_path=None, key=None, sigma=None, alpha=None,
-                log_name='pyfibre', database_name=None, debug=False,
+        with mock.patch(
+                'pyfibre.cli.pyfibre_app'
+                '.PyFibreApplication._run_pyfibre') as mock_pyfibre:
+            mock_pyfibre.side_effect = mock_run_pyfibre
+            pyfibre.cli.pyfibre_cli.run(
+                file_path='', key='', sigma=0.5, alpha=0.5,
+                log_name='pyfibre', database_name='', debug=False,
                 profile=False, ow_metric=False, ow_segment=False,
                 ow_network=False, save_figures=False, test=False
             )
-            self.log = pyfibre.cli.app.logging.getLogger(__name__)
+            self.log = pyfibre.cli.pyfibre_cli.logging.getLogger(__name__)
             # This test seems to be broken at the moment
             # self.assertEqual(10, self.log.getEffectiveLevel())
             if os.path.exists('pyfibre.log'):
                 os.remove('pyfibre.log')
 
     def test_run_with_profile(self):
-        with mock.patch('pyfibre.cli.app.PyFibreCLI') as mock_pyfibre:
-            mock_pyfibre.return_value = DummyPyFibreCLI()
-            pyfibre.cli.app.run(
-                file_path=None, key=None, sigma=None, alpha=None,
-                log_name='pyfibre', database_name=None, debug=False,
+        with mock.patch(
+                'pyfibre.cli.pyfibre_app'
+                '.PyFibreApplication._run_pyfibre') as mock_pyfibre:
+            mock_pyfibre.side_effect = mock_run_pyfibre
+            pyfibre.cli.pyfibre_cli.run(
+                file_path='', key='', sigma=0.5, alpha=0.5,
+                log_name='pyfibre', database_name='', debug=False,
                 profile=True, ow_metric=False, ow_segment=False,
                 ow_network=False, save_figures=False, test=False
             )
