+++ pyfibre/cli/tests/test_pyfibre_app.py
@@ -5,13 +5,12 @@ from tempfile import NamedTemporaryFile
 import pandas as pd
 
 from pyfibre.io.shg_pl_reader import SHGPLTransReader
-from pyfibre.pyfibre_runner import PyFibreRunner
 from pyfibre.tests.fixtures import test_shg_pl_trans_image_path
 
-from ..pyfibre_cli import PyFibreCLI
+from ..pyfibre_app import PyFibreApplication
 
 
-ITERATOR_PATH = 'pyfibre.cli.pyfibre_cli.analysis_generator'
+ITERATOR_PATH = 'pyfibre.cli.pyfibre_app.analysis_generator'
 
 
 def dummy_iterate_images(dictionary, runner, analysers, reader):
@@ -25,24 +24,22 @@ def dummy_iterate_images(dictionary, runner, analysers, reader):
         ]
 
 
-class TestPyFibreCLI(TestCase):
+class TestPyFibreApplication(TestCase):
 
     def setUp(self):
 
-        self.pyfibre_cli = PyFibreCLI()
+        self.pyfibre_app = PyFibreApplication()
 
     def test_init(self):
 
+        self.assertEqual(1, len(self.pyfibre_app.supported_readers))
         self.assertIsInstance(
-            self.pyfibre_cli.runner, PyFibreRunner)
-        self.assertEqual(1, len(self.pyfibre_cli.supported_readers))
-        self.assertIsInstance(
-            self.pyfibre_cli.supported_readers['SHG-PL-Trans'],
+            self.pyfibre_app.supported_readers['SHG-PL-Trans'],
             SHGPLTransReader)
 
     def test_init_pyfibre_workflow(self):
 
-        workflow = self.pyfibre_cli.runner
+        workflow = self.pyfibre_app.runner
 
         self.assertEqual(0.5, workflow.sigma)
         self.assertEqual(0.5, workflow.alpha)
@@ -53,13 +50,15 @@ class TestPyFibreCLI(TestCase):
 
     def test_save_database(self):
 
+        self.pyfibre_app.file_paths = [test_shg_pl_trans_image_path]
+
         with NamedTemporaryFile() as tmp_file:
-            self.pyfibre_cli.database_name = tmp_file.name
+            self.pyfibre_app.database_name = tmp_file.name
 
             with mock.patch(ITERATOR_PATH) as mock_iterate:
                 mock_iterate.side_effect = dummy_iterate_images
 
-                self.pyfibre_cli.run(test_shg_pl_trans_image_path)
+                self.pyfibre_app.run()
                 self.assertTrue(mock_iterate.called)
 
             self.assertTrue(
