+++ setup.py
@@ -34,7 +34,7 @@ setup(
     long_description=README_TEXT,
     packages=find_packages(),
     entry_points={
-        'gui_scripts': ['PyFibre = pyfibre.cli.app:pyfibre',
+        'gui_scripts': ['PyFibre = pyfibre.cli.pyfibre_cli:pyfibre',
                         'PyFibre_GUI = pyfibre.gui.app:pyfibre']
     },
     install_requires=REQUIREMENTS
